+++ Observer/Cancelorder.php
@@ -25,6 +25,11 @@ class Cancelorder implements \Magento\Framework\Event\ObserverInterface
     protected $reepayHelper;
 
     /**
+     * @var \Radarsofthouse\Reepay\Api\SessionRepositoryInterface
+     */
+    protected $reepaySessionRepository;
+
+    /**
      * @var \Radarsofthouse\Reepay\Helper\Logger
      */
     protected $logger;
@@ -36,6 +41,7 @@ class Cancelorder implements \Magento\Framework\Event\ObserverInterface
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
      * @param \Radarsofthouse\Reepay\Helper\Session $reepaySession
      * @param \Radarsofthouse\Reepay\Helper\Data $reepayHelper
+     * @param \Radarsofthouse\Reepay\Api\SessionRepositoryInterface $reepaySessionRepository
      * @param \Radarsofthouse\Reepay\Helper\Logger $logger
      */
     public function __construct(
@@ -43,12 +49,14 @@ class Cancelorder implements \Magento\Framework\Event\ObserverInterface
         \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
         \Radarsofthouse\Reepay\Helper\Session $reepaySession,
         \Radarsofthouse\Reepay\Helper\Data $reepayHelper,
+        \Radarsofthouse\Reepay\Api\SessionRepositoryInterface $reepaySessionRepository,
         \Radarsofthouse\Reepay\Helper\Logger $logger
     ) {
         $this->reepayCharge = $reepayCharge;
         $this->scopeConfig = $scopeConfig;
         $this->reepaySession = $reepaySession;
         $this->reepayHelper = $reepayHelper;
+        $this->reepaySessionRepository = $reepaySessionRepository;
         $this->logger = $logger;
     }
 
@@ -80,6 +88,21 @@ class Cancelorder implements \Magento\Framework\Event\ObserverInterface
                 );
 
                 $this->logger->addDebug("Delete charge for order #".$order->getIncrementId());
+            } elseif ($charge !== false && $charge['state'] == 'failed') {
+                try {
+                    $reepaySessions = $this->reepaySessionRepository->getListByOrderNumber($order->getIncrementId());
+                    if($reepaySessions->getTotalCount() > 0){
+                        /** @var \Radarsofthouse\Reepay\Api\Data\SessionInterface $reepaySession */
+                        foreach ($reepaySessions->getItems() as $reepaySession) {
+                            $this->reepaySession->delete($apiKey, $reepaySession->getHandle());
+                            $this->logger->addDebug("Delete session # {$reepaySession->getHandle()} for order #".$order->getIncrementId());
+                        }
+                    }
+                } catch (\Magento\Framework\Exception\LocalizedException | \Exception $e) {
+                    $this->logger->addError("Delete session for order #".$order->getIncrementId() . ' has errors  '. $e->getMessage());
+                }
+
+                $this->logger->addDebug("Delete session for order #".$order->getIncrementId());
             } else {
                 $cancelRes = $this->reepayCharge->cancel(
                     $apiKey,
@@ -87,12 +110,12 @@ class Cancelorder implements \Magento\Framework\Event\ObserverInterface
                 );
 
                 $this->logger->addDebug("Cancel charge for order #".$order->getIncrementId());
-    
+
                 if (!empty($cancelRes)) {
                     if ($cancelRes['state'] == 'cancelled') {
                         $_payment = $order->getPayment();
                         $this->reepayHelper->setReepayPaymentState($_payment, 'cancelled');
-    
+
                         // delete reepay session
                         $sessionRes = $this->reepaySession->delete(
                             $apiKey,